fix: BitReader 64 bit rotate and start game sync test

+++ src/lib/DeltaSnapshot/Pack/EntityCountReader.cs
@@ -29,7 +29,7 @@ namespace Piot.Surge.DeltaSnapshot.Pack
                 throw new Exception($"suspicious count {count}");
             }
 
-            return (uint)count;
+            return count;
         }
     }
 }
